moved AnimTriggerMessage to messaging
