+++ QSB/Messaging/AnimTriggerMessage.cs
@@ -1,26 +1,23 @@
 ﻿using QSB.Messaging;
 using UnityEngine.Networking;
 
-namespace QSB.Animation
+namespace QSB.Messaging
 {
-    public class AnimTriggerMessage : MessageBase
+    public class AnimTriggerMessage : PlayerMessage
     {
         public short TriggerId;
-        public uint SenderId;
         public float Value;
 
         public override void Deserialize(NetworkReader reader)
         {
             Value = reader.ReadSingle();
             TriggerId = reader.ReadInt16();
-            SenderId = reader.ReadPackedUInt32();
         }
 
         public override void Serialize(NetworkWriter writer)
         {
             writer.Write(Value);
             writer.Write(TriggerId);
-            writer.Write(SenderId);
         }
     }
 }
